+++ newrelic/src/main/scala/org/broadinstitute/dsde/workbench/newrelic/NewRelicMetricsInterpreter.scala
@@ -2,31 +2,32 @@ package org.broadinstitute.dsde.workbench.newrelic
 
 import java.util.concurrent.TimeUnit
 
-import cats.effect.{IO, Timer}
+import cats.ApplicativeError
+import cats.effect.{Async, Timer}
 import cats.implicits._
 import com.newrelic.api.agent.NewRelic
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, Future}
 
-class NewRelicMetricsInterpreter(appName: String) extends NewRelicMetrics {
+class NewRelicMetricsInterpreter[F[_]: Async](appName: String) extends NewRelicMetrics[F] {
   val prefix = s"Custom/$appName"
 
-  def timeIO[A](name: String, reportError: Boolean = false)(ioa: IO[A])(implicit timer: Timer[IO]): IO[A] =
+  def timeIO[A](name: String, reportError: Boolean = false)(fa: F[A])(implicit timer: Timer[F], ae: ApplicativeError[F, Throwable]): F[A] =
     for {
       start <- timer.clock.monotonic(TimeUnit.MILLISECONDS)
-      attemptedResult <- ioa.attempt
+      attemptedResult <- fa.attempt(ae)
       end <- timer.clock.monotonic(TimeUnit.MILLISECONDS)
       duration = end - start
       _ <- attemptedResult match {
         case Left(e) =>
           for {
-            _ <- IO(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}/failure", duration))
-            _ <- if(reportError) IO(NewRelic.noticeError(e)) else IO.unit
+            _ <- Async[F].delay(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}/failure", duration))
+            _ <- if(reportError) Async[F].delay[Unit](NewRelic.noticeError(e)) else Async[F].unit
           } yield ()
-        case Right(_) => IO(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}/success", duration))
+        case Right(_) => Async[F].delay(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}/success", duration))
       }
-      res <- IO.fromEither(attemptedResult)
+      res <- Async[F].fromEither(attemptedResult)
     } yield res
 
   def timeFuture[A](name: String, reportError: Boolean = false)(futureA: => Future[A])(implicit ec: ExecutionContext): Future[A] =
@@ -46,13 +47,13 @@ class NewRelicMetricsInterpreter(appName: String) extends NewRelicMetrics {
       res <- Future.fromTry(attemptedResult.toTry)
     } yield res
 
-  def gauge[A](name: String, value: Float): IO[Unit] = IO(NewRelic.recordMetric(s"${prefix}/${name}", value))
+  def gauge[A](name: String, value: Float): F[Unit] = Async[F].delay(NewRelic.recordMetric(s"${prefix}/${name}", value))
 
-  def incrementCounterIO[A](name: String, count: Int = 1): IO[Unit] = IO(NewRelic.incrementCounter(s"${prefix}/${name}", count))
+  def incrementCounter[A](name: String, count: Int = 1): F[Unit] = Async[F].delay(NewRelic.incrementCounter(s"${prefix}/${name}", count))
 
   def incrementCounterFuture[A](name: String, count: Int = 1)(implicit ec: ExecutionContext): Future[Unit] = Future(NewRelic.incrementCounter(s"${prefix}/${name}", count))
 
-  def recordResponseTimeIO(name: String, duration: Duration): IO[Unit] = IO(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}", duration.toMillis))
+  def recordResponseTime(name: String, duration: Duration): F[Unit] = Async[F].delay(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}", duration.toMillis))
 
   def recordResponseTimeFuture(name: String, duration: Duration)(implicit ec: ExecutionContext) = Future(NewRelic.recordResponseTimeMetric(s"${prefix}/${name}", duration.toMillis))
 }