+++ project/Dependencies.scala
@@ -7,8 +7,8 @@ object Dependencies {
   val googleV       = "1.22.0"
   val scalaLoggingV = "3.7.2"
   val scalaTestV    = "3.0.1"
-  val circeVersion = "0.11.1"
-  val http4sVersion = "0.20.3"
+  val circeVersion = "0.12.2"
+  val http4sVersion = "0.21.0-M5"
 
   def excludeGuavaJDK5(m: ModuleID): ModuleID = m.exclude("com.google.guava", "guava-jdk5")
 
@@ -28,7 +28,7 @@ object Dependencies {
 
   val selenium: ModuleID = "org.seleniumhq.selenium" % "selenium-java" % "3.11.0" % "test"
 
-  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "1.3.1"
+  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "2.0.0"
 
   // metrics-scala transitively pulls in io.dropwizard.metrics:metrics-core
   val metricsScala: ModuleID =      "nl.grons"              %% "metrics-scala"    % "3.5.6"
@@ -50,25 +50,28 @@ object Dependencies {
   val googleBigQuery: ModuleID =             "com.google.apis"       % "google-api-services-bigquery"             % s"v2-rev377-$googleV"
   val googleGuava: ModuleID = "com.google.guava"  % "guava" % "22.0"
   val googleRpc: ModuleID =               "io.grpc" % "grpc-core" % "1.16.1" //old google libraries relies on older version of grpc
-  val googleRpc2: ModuleID =               "io.grpc" % "grpc-core" % "1.16.1" //google2 may depends on newer version of grpc
+
+  val googleRpc2: ModuleID =               "io.grpc" % "grpc-core" % "1.24.0" //google2 may depends on newer version of grpc
   val googleFirestore: ModuleID = "com.google.cloud" % "google-cloud-firestore" % "0.71.0-beta"
-  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.77.0"
-  val googleStorageLocal: ModuleID = "com.google.cloud" % "google-cloud-nio" % "0.95.0-alpha" % "test"
+  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.93.0"
+  val googleStorageLocal: ModuleID = "com.google.cloud" % "google-cloud-nio" % "0.111.0-alpha" % "test"
   val googlePubsubNew: ModuleID = "com.google.cloud" % "google-cloud-pubsub" % "1.62.0"
   val googleKms: ModuleID = "com.google.cloud" % "google-cloud-kms" % "0.77.0-beta"
+  val googleDataproc: ModuleID =    "com.google.cloud" % "google-cloud-dataproc" % "0.111.0"
 
   val circeCore: ModuleID = "io.circe" %% "circe-core" % circeVersion
   val circeParser: ModuleID = "io.circe" %% "circe-parser" % circeVersion
   val circeGeneric: ModuleID = "io.circe" %% "circe-generic" % circeVersion % "test"
-  val circeFs2: ModuleID = "io.circe" %% "circe-fs2" % "0.11.0"
-  val log4cats = "io.chrisdavenport" %% "log4cats-slf4j"   % "0.4.0-M1"
+  val circeFs2: ModuleID = "io.circe" %% "circe-fs2" % "0.12.0"
+  val cats: ModuleID = "org.typelevel" %% "cats-core" % "2.0.0"
+  val log4cats = "io.chrisdavenport" %% "log4cats-slf4j"   % "1.0.0"
+  val catsMtl = "org.typelevel" %% "cats-mtl-core" % "0.7.0"
 
   val http4sCirce = "org.http4s" %% "http4s-circe" % http4sVersion
   val http4sBlazeClient = "org.http4s" %% "http4s-blaze-client" % http4sVersion
   val http4sDsl = "org.http4s"      %% "http4s-dsl"          % http4sVersion
 
-  val fs2Io: ModuleID = "co.fs2" %% "fs2-io" % "1.0.4"
-  val lineBacker: ModuleID = "io.chrisdavenport" %% "linebacker" % "0.2.1"
+  val fs2Io: ModuleID = "co.fs2" %% "fs2-io" % "2.0.1"
   val rawlsModel: ModuleID = "org.broadinstitute.dsde" %% "rawls-model" % "0.1-0d02c8ce-SNAP" exclude("com.typesafe.scala-logging", "scala-logging_2.11") exclude("com.typesafe.akka", "akka-stream_2.11")
   val newRelic: ModuleID = "com.newrelic.agent.java" % "newrelic-api" % "5.0.0"
 
@@ -87,12 +90,7 @@ object Dependencies {
     akkaHttpSprayJson,
     akkaTestkit,
     mockito,
-    log4cats,
-    circeCore,
-    circeParser,
-    circeGeneric,
-    fs2Io,
-    circeFs2
+    cats
   )
 
   val modelDependencies = commonDependencies ++ Seq(
@@ -143,18 +141,29 @@ object Dependencies {
     googleStorageLocal,
     googlePubsubNew,
     googleKms,
+    googleDataproc,
     http4sCirce,
     http4sBlazeClient,
     http4sDsl,
     log4cats,
     circeFs2,
-    lineBacker
+    catsMtl
   )
 
   val newrelicDependencies = Seq(
     catsEffect,
     log4cats,
     newRelic
+  )
+
+  val util2Dependencies = commonDependencies ++ List(
+    catsEffect,
+    log4cats,
+    fs2Io,
+    circeFs2,
+    circeCore,
+    circeParser,
+    circeGeneric
   )
 
   val serviceTestDependencies = commonDependencies ++ Seq(