+++ util2/src/test/scala/org/broadinstitute/dsde/workbench/util/WorkbenchTestSuite.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.workbench.util
+package org.broadinstitute.dsde.workbench.util2
 
 import cats.effect.{ContextShift, IO, Timer}
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
@@ -8,7 +8,7 @@ import org.scalatest.prop.{Configuration, PropertyChecks}
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.global
 
-trait WorkbenchTest {
+trait WorkbenchTestSuite {
   implicit val timer: Timer[IO] = IO.timer(global)
   implicit val cs: ContextShift[IO] = IO.contextShift(global)
   implicit val logger = Slf4jLogger.getLogger[IO]