+++ util/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file documents changes to the `workbench-util` library, including notes on how to upgrade to new versions.
 
+## 0.6
+### Changed
+- Moved code that depends on `circe`, `fs2` to `util2` since these libraries no longer support 2.11
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.6-TRAVIS-REPLACE-ME"`
+
 ## 0.5
 - add retry for IO
 - upgrade cats-effect